Removed useless square-rooting
Calculating sqrt is an operation that requires a lot of resources. As Double.compare only compares two numbers, we don‘t need to calculate the sqrt, as the square of whatever that number is is always bigger as well. 

(example)
sqrt(4) > sqrt(0) 
can be converted into
4 > 9

in other words,
if sqrt(x) > sqrt(y), then x > y
if sqrt(x) < sqrt(y), then x < y
if sqrt(x) == sqrt(y), then x == y

So, dont sqrt if it isnt needed, as computers need to calculate a lot