Fix issue with tab list size on low pop servers
I found a problem with my implementation which caused the tab list to take up more space than needed when the max size option was set to a value larger than the amount of players on your server.

This is fixed by using the already-injected collectPlayerEntries() method, instead of referencing the module setting directly.